feat(DAY-2): Add router, containers
1. Add router (react-router-native)
2. Add containers (Home, Map)
