+++ package.json
@@ -12,11 +12,13 @@
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
-    "react-native-web": "^0.11.7"
+    "react-native-web": "^0.11.7",
+    "react-router-native": "^5.1.2"
   },
   "devDependencies": {
     "@types/react": "^16.8.23",
     "@types/react-native": "^0.57.65",
+    "@types/react-router-native": "^5.1.0",
     "babel-preset-expo": "^7.0.0",
     "typescript": "^3.4.5"
   },
