+++ src/containers/Home.tsx
@@ -1,0 +1,23 @@
+import * as React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import { Link } from 'react-router-native';
+
+const Home = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Open up App.tsx to start working on your app!</Text>
+      <Link to='/map'>Move to Map</Link>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default Home;
